requirements.txt
# HTTP 요청
requests>=2.31.0
urllib3>=2.0.0

# 데이터베이스 연결
pyodbc>=5.0.0
vertica-python>=1.3.0

# CLI 인터페이스
click>=8.1.0

# 설정 관리
pyyaml>=6.0
python-dotenv>=1.0.0

# SSH 연결
paramiko>=3.0.0

# Java 연동 (DataStage SDK)
JPype1>=1.4.0

# 로깅 (표준 라이브러리 포함)

